[MIG] pos_payment_method_cashdro: migration to 16.0
TT45180
