+++ pos_payment_method_cashdro/i18n/pos_payment_method_cashdro.pot
@@ -4,7 +4,7 @@
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 14.0\n"
+"Project-Id-Version: Odoo Server 16.0\n"
 "Report-Msgid-Bugs-To: \n"
 "Last-Translator: \n"
 "Language-Team: \n"
@@ -29,26 +29,16 @@ msgid "Cashdro User"
 msgstr ""
 
 #. module: pos_payment_method_cashdro
-#: model:ir.model.fields,field_description:pos_payment_method_cashdro.field_pos_payment_method__display_name
-msgid "Display Name"
-msgstr ""
-
-#. module: pos_payment_method_cashdro
-#: model:ir.model.fields,field_description:pos_payment_method_cashdro.field_pos_payment_method__id
-msgid "ID"
-msgstr ""
-
-#. module: pos_payment_method_cashdro
 #: model:ir.model.fields,help:pos_payment_method_cashdro.field_pos_payment_method__cashdro_host
 msgid "It must be reachable by the PoS in the store"
 msgstr ""
 
 #. module: pos_payment_method_cashdro
-#: model:ir.model.fields,field_description:pos_payment_method_cashdro.field_pos_payment_method____last_update
-msgid "Last Modified on"
+#: model:ir.model,name:pos_payment_method_cashdro.model_pos_payment_method
+msgid "Point of Sale Payment Methods"
 msgstr ""
 
 #. module: pos_payment_method_cashdro
-#: model:ir.model,name:pos_payment_method_cashdro.model_pos_payment_method
-msgid "Point of Sale Payment Methods"
+#: model:ir.model,name:pos_payment_method_cashdro.model_pos_session
+msgid "Point of Sale Session"
 msgstr ""
