+++ pos_payment_method_cashdro/models/__init__.py
@@ -1,1 +1,2 @@
 from . import pos_payment_method
+from . import pos_session
