+++ pos_payment_method_cashdro/models/pos_session.py
@@ -1,0 +1,14 @@
+# Copyright 2024 Tecnativa - David Vidal
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
+from odoo import models
+
+
+class PosSession(models.Model):
+    _inherit = "pos.session"
+
+    def _loader_params_pos_payment_method(self):
+        result = super()._loader_params_pos_payment_method()
+        result["search_params"]["fields"].extend(
+            ["cashdro_host", "cashdro_user", "cashdro_password"]
+        )
+        return result
