--- pos_payment_method_cashdro/views/assets.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <template id="assets" inherit_id="point_of_sale.assets">
-        <xpath expr="." position="inside">
-            <script
-                type="text/javascript"
-                src="/pos_payment_method_cashdro/static/src/js/models.js"
-            />
-            <script
-                type="text/javascript"
-                src="/pos_payment_method_cashdro/static/src/js/payment_cashdro.js"
-            />
-      </xpath>
-    </template>
-</odoo>
